feat: added mul2 div2 get, set and clear bit

+++ include/bitwise.h
@@ -5,11 +5,11 @@
 #include <sys/types.h>
 
 uint8_t is_even(int n);
-uint8_t mul2(int n);
-uint8_t div2(int n);
+int mul2(int n);
+int div2(int n);
 
-uint8_t get_bit(int n, int pos);
-uint8_t set_bit(int n, int pos);
+int get_bit(int n, int pos);
+int set_bit(int n, int pos);
 uint8_t clear_bit(int n, int pos);
 
 uint8_t popcount(uint n);
